// backups/dashboard-ui/20260610_115652/src/types/api.ts
export interface SlideItem {
  internal_id: number
  global_artifact_id: string | null
  original_filename: string | null
  current_filename: string | null
  status: string | null
  file_size: number | null
  file_format: string | null
  scanner_id: string | null
  scanner_name: string | null
  artifact_type: string | null
  created_at: string | null
  updated_at: string | null
  original_path: string | null
  current_file_path: string | null
}